requirements.txt
streamlit>=1.18.0
langchain>=0.0.300
pydantic>=1.10.0
python-dotenv>=1.0.0
openai>=0.27.0
flake8>=6.0.0
pytest>=7.0.0
httpx>=0.23.0
tqdm>=4.0.0
requests>=2.28.0
